Issue 3050 - Fix broken AA links on google.com
This is most definitely a temporary workaround. In this particular
case the very generous whitelisting rule is OK, since there's nothing
that's not an Acceptable Ad on google.com anyway. But we have to find a better
workaround to replace the specific exception rules not working here. And we
surely shouldn't have to manually modify the rules, this needs to be covered
in the conversion script.